refactor: consolidate git/file operations into shared utilities
Created src/utils/git/exec.ts (git, execSafe) and src/utils/fs.ts
(ensureDir, fileExists, readFile, writeFile). Moved gh/ghRaw to
gh-cli-wrapper.ts. Updated all consumers.
