+++ src/lib/auto-claude/index.ts
@@ -1,7 +1,8 @@
 export { type AutoClaudeConfig, AutoClaudeConfigSchema, getConfig, initConfig } from "./config.js";
 export { STEP_NAMES, runPipeline } from "./pipeline.js";
 export type { StepName } from "./prompt-templates/index.js";
-export { git, sleep } from "./shell.js";
+export { git } from "../../utils/git/exec.js";
+export { sleep } from "./shell.js";
 export { fetchIssue, fetchIssues } from "./steps/fetch-issues.js";
 export {
   type IssueContext,
