+++ src/lib/auto-claude/pipeline.ts
@@ -10,8 +10,10 @@ import { stepPlan } from "./steps/plan.js";
 import { stepReview } from "./steps/review.js";
 import { stepSimplify } from "./steps/simplify.js";
 import { LABELS, ensureLabelsExist, removeLabel, setLabel } from "./labels.js";
-import { execSafe, ghRaw, git } from "./shell.js";
-import { ensureDir, fileExists, log, readFile, writeFile } from "./utils.js";
+import { ensureDir, fileExists, readFile, writeFile } from "../../utils/fs.js";
+import { execSafe, git } from "../../utils/git/exec.js";
+import { ghRaw } from "../../utils/git/gh-cli-wrapper.js";
+import { log } from "./utils.js";
 import type { IssueContext } from "./utils.js";
 
 export { type StepName, STEP_NAMES } from "./prompt-templates/index.js";
