+++ src/utils/fs.ts
@@ -1,0 +1,19 @@
+import { existsSync, mkdirSync, readFileSync, writeFileSync } from "node:fs";
+import { dirname } from "node:path";
+
+export function ensureDir(dir: string): void {
+  mkdirSync(dir, { recursive: true });
+}
+
+export function fileExists(path: string): boolean {
+  return existsSync(path);
+}
+
+export function readFile(path: string): string {
+  return readFileSync(path, "utf-8");
+}
+
+export function writeFile(path: string, content: string): void {
+  ensureDir(dirname(path));
+  writeFileSync(path, content, "utf-8");
+}
